[WIP] rework the VME interface
Attempt to be more OO in the interface.
Methods take inputs as arguments and return ouputs. Everything is
strongly typed through templating.

This should also ease the python interface.

All this is work in progress... I modified the general classes, but now
all VME board implementations have to be adapted, starting with the
CAENET adapter.